Fixed issue 2
